+++ contracts/Swap.sol
@@ -24,7 +24,7 @@ contract Swap is AbstractSwap, Initializable {
         return success ? abi.decode(result, (int32)) : HederaResponseCodes.UNKNOWN;
     }
 
-    function transferToken(address _token, address sender, address receiver, int64 amount) internal override virtual returns(int) {
+    function transferToken(address _token, address sender, address receiver, int amount) internal override virtual returns(int) {
         (bool success, bytes memory result) = address(tokenService).delegatecall(
             abi.encodeWithSelector(IBaseHTS.transferTokenPublic.selector,
             _token, sender, receiver, amount));
