+++ contracts/test/LPTokenTest.sol
@@ -13,9 +13,9 @@ contract LPTokenTest is AbstractLPToken {
          tokenService = _tokenService;
     }
 
-    function mintToken(uint64 amount) override internal virtual 
-     returns (int responseCode, uint64 newTotalSupply) {
-            (int response, uint64 _newTotalSupply) = tokenService.mintTokenPublic(address(lpToken), amount);
+    function mintToken(int amount) override internal virtual 
+     returns (int responseCode, int newTotalSupply) {
+            (int response, int _newTotalSupply) = tokenService.mintTokenPublic(address(lpToken), amount);
 
              if (response != HederaResponseCodes.SUCCESS) {
                 revert ("Mint Failed");
@@ -23,7 +23,7 @@ contract LPTokenTest is AbstractLPToken {
             return (response, _newTotalSupply);
     }
 
-    function burnToken(uint64 amount) override internal virtual returns (int) {
+    function burnToken(int amount) override internal virtual returns (int) {
         (int responseCode, ) = tokenService.burnTokenPublic(address(lpToken), amount);
 
         if (responseCode != HederaResponseCodes.SUCCESS) {
@@ -36,7 +36,7 @@ contract LPTokenTest is AbstractLPToken {
         return tokenService.associateTokenPublic(account, _token);
     }
 
-    function transferTokenInternal(address _token, address sender, address receiver, int64 amount) internal override virtual returns(int) {
+    function transferTokenInternal(address _token, address sender, address receiver, int amount) internal override virtual returns(int) {
         int responseCode = tokenService.transferTokenPublic(_token, sender, receiver, amount);
         if (responseCode != HederaResponseCodes.SUCCESS) {
                 revert ("LP Token Transfer Fail");
